Moved ApiError from client.ts to api.ts
