+++ ui/packages/api-client/src/api.ts
@@ -6379,6 +6379,47 @@ This struct contains optional fields that can be updated on a TextTranslation re
   },
 ] as const satisfies ZodiosEndpointDefinitions);
 
+// Axios error for failing request
+export interface ApiError {
+	// Standard
+	message: string;
+	name: string;
+	// Microsoft
+	description: string;
+	number: number;
+	// Mozilla
+	fileName: string;
+	lineNumber: number;
+	columnNumber: number;
+	stack: string;
+	// Axios
+	config: {
+		adapter: string[];
+		allowAbsoluteUrls: boolean;
+		baseURL:string;
+		data: undefined
+		env: object;
+		headers: object;
+		maxBodyLength: number;
+		maxContentLength: number;
+		method: string;
+		params: object;
+		timeout: number;
+		transformRequest: string[]
+		transformResponse: string[]
+		transitional: { silentJSONParsing: boolean, forcedJSONParsing: boolean, clarifyTimeoutError: boolean  }
+		url: string;
+		validateStatus: (status: string) => void;
+		withCredentials: true
+		xsrfCookieName: string;
+		xsrfHeaderName: string;
+	}
+	code: string;
+	status: number;
+	response: Response;
+	request: Request;
+}
+
 export const api: ZodiosInstance<typeof endpoints> = new Zodios(endpoints);
 export type ApiClient = typeof api;
